Move piped error processing to interceptor
